requirements.txt
Django==4.2.2
mysql-connector-python==8.0.33
mysqlclient==2.2.0
selenium==4.10.0
